fix: fix broken test in s2s bridge issuing
