+++ frame/evm/src/lib.rs
@@ -459,6 +459,15 @@ pub trait BlockHashMapping {
 	fn block_hash(number: u32) -> H256;
 }
 
+/// Returns the Substrate block hash by number.
+pub struct SubstrateBlockHashMapping<T>(sp_std::marker::PhantomData<T>);
+impl<T: Config> BlockHashMapping for SubstrateBlockHashMapping<T> {
+	fn block_hash(number: u32) -> H256 {
+		let number = T::BlockNumber::from(number);
+		H256::from_slice(frame_system::Pallet::<T>::block_hash(number).as_ref())
+	}
+}
+
 /// A contract handle for ethereum issuing
 pub trait IssuingHandler {
 	fn handle(address: H160, caller: H160, input: &[u8]) -> DispatchResult;
